requirements.txt
discord.py

coloredlogs
watchdog
pyyaml
timedelta
python-dotenv

aiohttp

redis
motor